reorganizacion
se organizaron los archivos y las rutas

+++ backend/src/services/menuServices.js
@@ -1,48 +1,44 @@
+const { ExistNameCategoriaError, NotFoundError } = require('../errors/ExceptionErrors');
 const menuRepository = require('../repositories/menuRepository');
 
 
 //Servicio para mostrar las categorias
 exports.mostrarCategorias = async () => {
-    try {
-        return await menuRepository.mostrarCategorias();
-    } catch (error) {
-        throw new Error(error);
-    }
+    return await menuRepository.mostrarCategorias();
 };
 
 //Servicio para mostrar una categoria
 exports.mostrarCategoria = async (id) => {
-    try {
-        return await menuRepository.mostrarCategoria(id);
-    } catch (error) {
-        throw new Error(error);
-    }
+    const existe = await menuRepository.mostrarCategoria(id);
+    if(existe.length <= 0) throw new NotFoundError('La categoria no existe');
+    return existe
 };
 
 //Servicio para crear una categoria
 exports.crearCategoria = async (categoria, foto) => {
-    try {
-        const response = await menuRepository.crearCategoria(categoria, foto);
-        return response;
-    } catch (error) {
-        throw new Error(error);
-    }
+    //Verificar si la categoria ya existe
+    const existe = await menuRepository.verificarNombre(categoria);
+    if (existe) throw new ExistNameCategoriaError();
+
+    const response = await menuRepository.crearCategoria(categoria, foto);
+    return response;
 };
 
 //Servicio para actualizar una categoria
 exports.actualizarCategoria = async (id, foto, categoria) => {
-    try {
-        const reponse = await menuRepository.actualizarCategoria(id, foto, categoria);
-        return reponse
-    } catch (error) {
-        throw new Error(error);
-    }
+    const existe = await this.mostrarCategoria(id)
+    const existe2 = await menuRepository.verificarNombre(categoria);
+
+    if (existe.length <= 0) throw new NotFoundError('la categoria no existe');
+    if (existe2) throw new ExistNameCategoriaError();
+    const reponse = await menuRepository.actualizarCategoria(id, foto, categoria);
+    return reponse
 }
 
 exports.eliminarCategoria = async (id) => {
-    try {
-        return await menuRepository.eliminarCategoria(id);
-    } catch (error) {
-        throw new Error(error);
-    }
+    const existe = await this.mostrarCategoria(id)
+
+    if (existe.length <= 0) throw new NotFoundError('la categoria no existe');
+    const response = await menuRepository.eliminarCategoria(id);
+    return response
 }
